refactor: move MessageType to constants, removing console.log and minor changes
